Added Instagram

+++ wp-content/plugins/vd-socializer/inc/Assets/database_settings_template.php
@@ -1,33 +1,41 @@
 <?php
 
 use Inc\Admin\AdminMenu;
-if($_POST['vd_hidden'] == 'Y') {
+if(key_exists('vd_hidden',$_POST)) {
+	if ( $_POST['vd_hidden'] == 'Y' ) {
 
-    $facebook_app = $_POST['vd_facebook_app'];
-    update_option('vd_facebook_app', $facebook_app);
+		$facebook_app = $_POST['vd_facebook_app'];
+		update_option( 'vd_facebook_app', $facebook_app );
 
-    $facebook_secret = $_POST['vd_facebook_secret'];
-    update_option('vd_facebook_secret', $facebook_secret);
+		$facebook_secret = $_POST['vd_facebook_secret'];
+		update_option( 'vd_facebook_secret', $facebook_secret );
 
-	$twitter_app = $_POST['vd_twitter_app'];
-	update_option('vd_twitter_app', $twitter_app);
+		$twitter_app = $_POST['vd_twitter_app'];
+		update_option( 'vd_twitter_app', $twitter_app );
 
-	$twitter_secret = $_POST['vd_twitter_secret'];
-	update_option('vd_twitter_secret', $twitter_secret);
+		$twitter_secret = $_POST['vd_twitter_secret'];
+		update_option( 'vd_twitter_secret', $twitter_secret );
 
-        ?>
-        <div class="updated"><p><strong><?php _e('Options saved.' ); ?></strong></p></div>
-    <?php
+		$instagram_app = $_POST['vd_instagram_app'];
+		update_option( 'vd_instagram_app', $instagram_app );
+
+		$instagram_secret = $_POST['vd_instagram_secret'];
+		update_option( 'vd_instagram_secret', $instagram_secret );
+
+		?>
+        <div class="updated"><p><strong><?php _e( 'Options saved.' ); ?></strong></p></div>
+		<?php
+	}
 }
-else{
-    $dbhost = get_option('vd_dbhost');
-    $dbname = get_option('vd_dbname');
-    $dbuser = get_option('vd_dbuser');
-    $dbpwd = get_option('vd_dbpwd');
-    $facebook_app = get_option('vd_facebook_app');
-    $facebook_secret = get_option('vd_facebook_secret');
-    $twitter_app = get_option('vd_twitter_app');
-    $twitter_secret = get_option('vd_twitter_secret');
+else {
+
+		$facebook_app    = get_option( 'vd_facebook_app' );
+		$facebook_secret = get_option( 'vd_facebook_secret' );
+		$twitter_app     = get_option( 'vd_twitter_app' );
+		$twitter_secret  = get_option( 'vd_twitter_secret' );
+		$instagram_app = get_option('vd_instagram_app');
+	    $instagram_secret = get_option('vd_instagram_secret');
+
 }
 
 
@@ -47,6 +55,10 @@ else{
 	    <?php    echo "<h4>" . __( 'Twitter Settings', 'vd_trdom' ) . "</h4>"; ?>
 	    <p><?php _e("Twitter App Id: " ); ?><input type="text" name="vd_twitter_app" value="<?php echo $twitter_app; ?>" size="50"><?php _e("ex: QJd21c6OVDluWLrTlWziAEtk8" ); ?></p>
 	    <p><?php _e("Twitter App Secret: " ); ?><input type="text" name="vd_twitter_secret" value="<?php echo $twitter_secret; ?>" size="100"><?php _e("ex: 8Dz6DmAxE38uoojW99qSNgbae8mmZWayvSTK9y4v2n9jmyb8rP" ); ?></p>
+
+	    <?php    echo "<h4>" . __( 'Instagram Settings', 'vd_trdom' ) . "</h4>"; ?>
+        <p><?php _e("Instagram App Id: " ); ?><input type="text" name="vd_instagram_app" value="<?php echo $instagram_app; ?>" size="50"><?php _e("ex: QJd21c6OVDluWLrTlWziAEtk8" ); ?></p>
+        <p><?php _e("Instagram App Secret: " ); ?><input type="text" name="vd_instagram_secret" value="<?php echo $instagram_secret; ?>" size="100"><?php _e("ex: 8Dz6DmAxE38uoojW99qSNgbae8mmZWayvSTK9y4v2n9jmyb8rP" ); ?></p>
 
         <p class="submit">
         <input type="submit" name="Submit" value="<?php _e('Update Options', 'vd_trdom' ) ?>" />
